+++ lib/src/helpers/workday_helper.dart
@@ -1,6 +1,6 @@
 import 'package:time/time.dart';
 
-import '../due_date.dart';
+import '../every.dart';
 
 /// Helper class to work with workdays.
 class WorkdayHelper {
@@ -48,15 +48,4 @@ class WorkdayHelper {
     if (isNext) return workdays.next(date);
     return workdays.previous(date);
   }
-}
-
-/// Extension for [Object]s.
-extension ObjectExt<T extends Object> on T {
-  /// Returns the result of [orElse] if [predicate] returns false, otherwise
-  /// returns this.
-  T? when(bool Function(T self) predicate, {T? Function(T self)? orElse}) =>
-      predicate(this) ? this : orElse?.call(this);
-
-  /// Applies [f] to this object and returns the result.
-  R apply<R>(R Function(T self) f) => f(this);
 }
