+++ test/helpers/workday_helper_test.dart
@@ -1,5 +1,5 @@
 import 'package:due_date/src/every.dart';
-import 'package:due_date/src/shared_private.dart';
+import 'package:due_date/src/helpers/workday_helper.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -79,67 +79,6 @@ void main() {
           WorkdayHelper.adjustToWorkday(date, isNext: false),
           equals(DateTime(2024, DateTime.november, 8)),
         );
-      });
-    });
-  });
-  group('ObjectExt', () {
-    group('when', () {
-      final object = Object();
-      test('predicate parameter is the same object as the caller', () {
-        object.when((self) {
-          expect(self, same(object));
-          return false;
-        });
-      });
-      test('orElse parameter is the same object as the caller', () {
-        object.when(
-          (_) => false,
-          orElse: (self) {
-            expect(self, same(object));
-            return false;
-          },
-        );
-      });
-      test('orElse is not called if predicate is true', () {
-        expect(
-          object.when(
-            (_) => true,
-            orElse: (_) {
-              throw AssertionError('orElse should not be called');
-            },
-          ),
-          equals(object),
-        );
-      });
-      test('predicate is true', () {
-        expect(object.when((_) => true), equals(object));
-      });
-      group('predicate is false', () {
-        test('orElse is null', () {
-          expect(object.when((_) => false), isNull);
-        });
-        group('orElse is not null', () {
-          test('orElse returns null', () {
-            expect(object.when((_) => false, orElse: (_) => null), isNull);
-          });
-          test('orElse returns value', () {
-            final newObject = Object();
-            expect(
-              object.when((_) => false, orElse: (_) => newObject),
-              equals(newObject),
-            );
-          });
-        });
-      });
-    });
-    group('apply', () {
-      final object = Object();
-      test('returns self, parameter is the same as object', () {
-        expect(object.apply((self) => self), equals(object));
-      });
-      test('returns new object', () {
-        final newObject = Object();
-        expect(object.apply((_) => newObject), equals(newObject));
       });
     });
   });
